fix public identity consistency
