+++ .github/CONTRIBUTING.md
@@ -14,11 +14,11 @@ certify that:
 - It is your original work, or you have permission to submit it.
 - It contains no secrets, private data, copied proprietary material, or
   conflicting license terms.
-- You grant PT. Nakafa Tekno Kreatif a perpetual, worldwide, non-exclusive,
+- You grant PT NAKAFA TEKNO KREATIF a perpetual, worldwide, non-exclusive,
   royalty-free, sublicensable, and transferable license to use, reproduce,
   modify, distribute, publicly display, publicly perform, create derivative
   works from, and relicense it as part of Nakafa.
-- PT. Nakafa Tekno Kreatif may use it in source-available, commercial,
+- PT NAKAFA TEKNO KREATIF may use it in source-available, commercial,
   proprietary, hosted, educational, and internal versions of Nakafa without
   owing you payment.
