+++ LICENSE
@@ -2,7 +2,7 @@
 
 Effective date: April 24, 2026
 
-Copyright (c) 2026 PT. Nakafa Tekno Kreatif. All rights reserved.
+Copyright (c) 2026 PT NAKAFA TEKNO KREATIF. All rights reserved.
 
 This is not an open-source license. Nakafa source code is public so people can
 inspect it, learn from it, review it, and propose improvements. The public
@@ -25,8 +25,8 @@ team, or business purpose.
 without making Nakafa available to any other person over the internet, an
 intranet, a private network, a shared server, or any hosted environment.
 
-`Written permission` means a separate written agreement signed by PT. Nakafa
-Tekno Kreatif, or a written approval sent from an official Nakafa email address.
+`Written permission` means a separate written agreement signed by PT NAKAFA
+TEKNO KREATIF, or a written approval sent from an official Nakafa email address.
 
 ## 2. Scope
 
@@ -47,13 +47,13 @@ This license does not apply to:
   received. New changes published after the effective date above are covered by
   this license.
 
-All rights not expressly granted by this license are reserved by PT. Nakafa Tekno
-Kreatif.
+All rights not expressly granted by this license are reserved by PT NAKAFA TEKNO
+KREATIF.
 
 ## 3. Allowed Use Without Written Permission
 
-You may do only the following without written permission from PT. Nakafa Tekno
-Kreatif:
+You may do only the following without written permission from PT NAKAFA TEKNO
+KREATIF:
 
 - View, read, inspect, and audit the source code.
 - Clone or download one copy for personal review.
@@ -72,7 +72,7 @@ revocable if you violate this license.
 
 ## 4. Permission Required
 
-You must get prior written permission from PT. Nakafa Tekno Kreatif before doing
+You must get prior written permission from PT NAKAFA TEKNO KREATIF before doing
 anything not listed in Section 3.
 
 The following uses always require prior written permission:
@@ -131,7 +131,7 @@ Your rights under this license end immediately if you violate any term.
 After termination, you must stop using the software and delete all copies under
 your control, except copies you are required to keep for legal compliance.
 
-PT. Nakafa Tekno Kreatif may reinstate your rights only in writing.
+PT NAKAFA TEKNO KREATIF may reinstate your rights only in writing.
 
 ## 9. Mandatory Legal Rights
 
@@ -143,13 +143,13 @@ limited, or conditioned by contract.
 Nakafa is provided "as is" and "as available", without warranties or conditions
 of any kind, whether express, implied, statutory, or otherwise.
 
-PT. Nakafa Tekno Kreatif does not warrant that Nakafa is accurate, secure,
+PT NAKAFA TEKNO KREATIF does not warrant that Nakafa is accurate, secure,
 complete, uninterrupted, error-free, fit for a particular purpose, or
 non-infringing.
 
 ## 11. Limitation of Liability
 
-To the fullest extent permitted by law, PT. Nakafa Tekno Kreatif will not be
+To the fullest extent permitted by law, PT NAKAFA TEKNO KREATIF will not be
 liable for any indirect, incidental, special, consequential, exemplary, punitive,
 or similar damages, or for any loss of profits, revenue, data, goodwill, or
 business opportunity, arising from or related to Nakafa or this license.
@@ -165,10 +165,10 @@ Republic of Indonesia.
 
 ## 13. No Waiver
 
-If PT. Nakafa Tekno Kreatif does not enforce a term immediately, that does not
+If PT NAKAFA TEKNO KREATIF does not enforce a term immediately, that does not
 waive the right to enforce that term later.
 
-Any waiver must be in writing and signed by PT. Nakafa Tekno Kreatif.
+Any waiver must be in writing and signed by PT NAKAFA TEKNO KREATIF.
 
 ## 14. Severability
