+++ apps/www/app/[locale]/(app)/(shared)/(site)/(legal)/privacy-policy/en.mdx
@@ -2,7 +2,7 @@
 
 Last updated: January 26, 2026
 
-This Privacy Policy explains how PT. Nakafa Tekno Kreatif ("Nakafa," "we," "us," or "our") collects, uses, and discloses Personal Data when you access or use our websites, applications, and related services (collectively, the "Services").
+This Privacy Policy explains how PT NAKAFA TEKNO KREATIF ("Nakafa," "we," "us," or "our") collects, uses, and discloses Personal Data when you access or use our websites, applications, and related services (collectively, the "Services").
 
 Your use of the Services is at all times subject to our [Terms of Service](/en/terms-of-service), which incorporates this Privacy Policy.
 
@@ -20,7 +20,7 @@ This Privacy Policy does not cover the practices of companies we do not own or c
 
 ## Important notice about third-party data processing
 
-This Privacy Policy covers how PT. Nakafa Tekno Kreatif acts as a data controller for Personal Data collected through the Services. However, we engage third-party service providers that act as data processors to provide core functionality, including hosting, database services, analytics, email delivery, subscription management, and AI processing.
+This Privacy Policy covers how PT NAKAFA TEKNO KREATIF acts as a data controller for Personal Data collected through the Services. However, we engage third-party service providers that act as data processors to provide core functionality, including hosting, database services, analytics, email delivery, subscription management, and AI processing.
 
 These third parties have their own privacy policies and data processing practices. We have entered into data processing agreements with these service providers that include data protection obligations and, where required, Standard Contractual Clauses for international data transfers. We only share Personal Data with these service providers as necessary to provide the Services and under contractually binding data protection obligations.
 
@@ -173,7 +173,6 @@ If you are making a GDPR-related request, include “GDPR Request” in the subj
 
 ## Contact information
 
-PT. Nakafa Tekno Kreatif  
-Taman Sukahati Permai H6  
-Kabupaten Bogor, Indonesia  
+PT NAKAFA TEKNO KREATIF, Taman Sukahati Permai Blok H-6, Sukahati, Cibinong, Kabupaten Bogor, Jawa Barat 16913, Indonesia
+
 Email: [nakafaai@gmail.com](mailto:nakafaai@gmail.com)
