+++ apps/www/app/[locale]/(app)/(shared)/(site)/(legal)/privacy-policy/id.mdx
@@ -2,7 +2,7 @@
 
 Terakhir diperbarui: 26 Januari 2026
 
-Kebijakan Privasi ini menjelaskan bagaimana PT. Nakafa Tekno Kreatif ("Nakafa", "kami") mengumpulkan, menggunakan, dan mengungkapkan Data Pribadi ketika Anda mengakses atau menggunakan situs web, aplikasi, dan layanan terkait kami (secara kolektif, "Layanan").
+Kebijakan Privasi ini menjelaskan bagaimana PT NAKAFA TEKNO KREATIF ("Nakafa", "kami") mengumpulkan, menggunakan, dan mengungkapkan Data Pribadi ketika Anda mengakses atau menggunakan situs web, aplikasi, dan layanan terkait kami (secara kolektif, "Layanan").
 
 Penggunaan Anda atas Layanan selalu tunduk pada [Syarat dan Ketentuan](/id/terms-of-service), yang memasukkan Kebijakan Privasi ini.
 
@@ -20,7 +20,7 @@ Kebijakan Privasi ini tidak mencakup praktik perusahaan yang tidak kami miliki a
 
 ## Pemberitahuan penting tentang pemrosesan data pihak ketiga
 
-Kebijakan Privasi ini mencakup bagaimana PT. Nakafa Tekno Kreatif bertindak sebagai pengendali data untuk Data Pribadi yang dikumpulkan melalui Layanan. Namun, kami bekerja sama dengan penyedia layanan pihak ketiga yang bertindak sebagai pemroses data untuk menyediakan fungsionalitas inti, termasuk hosting, layanan database, analitik, pengiriman email, manajemen langganan, dan pemrosesan AI.
+Kebijakan Privasi ini mencakup bagaimana PT NAKAFA TEKNO KREATIF bertindak sebagai pengendali data untuk Data Pribadi yang dikumpulkan melalui Layanan. Namun, kami bekerja sama dengan penyedia layanan pihak ketiga yang bertindak sebagai pemroses data untuk menyediakan fungsionalitas inti, termasuk hosting, layanan database, analitik, pengiriman email, manajemen langganan, dan pemrosesan AI.
 
 Pihak ketiga ini memiliki kebijakan privasi dan praktik pemrosesan data mereka sendiri. Kami telah mengadakan perjanjian pemrosesan data dengan penyedia layanan ini yang mencakup kewajiban perlindungan data dan, bila diperlukan, Standard Contractual Clauses untuk transfer data internasional. Kami hanya berbagi Data Pribadi dengan penyedia layanan ini bila diperlukan untuk menyediakan Layanan dan di bawah kewajiban perlindungan data yang mengikat secara kontraktual.
 
@@ -181,7 +181,6 @@ Jika Anda mengajukan permintaan terkait GDPR, sertakan \"GDPR Request\" pada sub
 
 ## Informasi kontak
 
-PT. Nakafa Tekno Kreatif  
-Taman Sukahati Permai H6  
-Kabupaten Bogor, Indonesia  
+PT NAKAFA TEKNO KREATIF, Taman Sukahati Permai Blok H-6, Sukahati, Cibinong, Kabupaten Bogor, Jawa Barat 16913, Indonesia
+
 Email: [nakafaai@gmail.com](mailto:nakafaai@gmail.com)
