+++ apps/www/app/[locale]/(app)/(shared)/(site)/(legal)/security-policy/en.mdx
@@ -2,7 +2,7 @@
 
 Last updated: January 26, 2026
 
-This Security Policy describes the security measures PT. Nakafa Tekno Kreatif ("Nakafa," "we," "us," or "our") uses, and the measures we expect users to take, to help protect the confidentiality, integrity, and availability of the Services and related data.
+This Security Policy describes the security measures PT NAKAFA TEKNO KREATIF ("Nakafa," "we," "us," or "our") uses, and the measures we expect users to take, to help protect the confidentiality, integrity, and availability of the Services and related data.
 
 For information about how we collect and use Personal Data, please review our [Privacy Policy](/en/privacy-policy).
 
@@ -103,7 +103,6 @@ We may update this Security Policy from time to time to reflect changes in our s
 
 ## Contact
 
-PT. Nakafa Tekno Kreatif  
-Taman Sukahati Permai H6  
-Kabupaten Bogor, Indonesia  
+PT NAKAFA TEKNO KREATIF, Taman Sukahati Permai Blok H-6, Sukahati, Cibinong, Kabupaten Bogor, Jawa Barat 16913, Indonesia
+
 Email: [nakafaai@gmail.com](mailto:nakafaai@gmail.com)
