+++ apps/www/app/[locale]/(app)/(shared)/(site)/(legal)/security-policy/id.mdx
@@ -2,7 +2,7 @@
 
 Terakhir diperbarui: 26 Januari 2026
 
-Kebijakan Keamanan ini menjelaskan langkah-langkah keamanan yang digunakan PT. Nakafa Tekno Kreatif ("Nakafa", "kami") serta langkah yang kami harapkan dari pengguna untuk membantu melindungi kerahasiaan, integritas, dan ketersediaan Layanan serta data terkait.
+Kebijakan Keamanan ini menjelaskan langkah-langkah keamanan yang digunakan PT NAKAFA TEKNO KREATIF ("Nakafa", "kami") serta langkah yang kami harapkan dari pengguna untuk membantu melindungi kerahasiaan, integritas, dan ketersediaan Layanan serta data terkait.
 
 Untuk informasi tentang cara kami mengumpulkan dan menggunakan Data Pribadi, silakan tinjau [Kebijakan Privasi](/id/privacy-policy).
 
@@ -103,7 +103,6 @@ Kami dapat memperbarui Kebijakan Keamanan ini dari waktu ke waktu untuk mencermi
 
 ## Kontak
 
-PT. Nakafa Tekno Kreatif  
-Taman Sukahati Permai H6  
-Kabupaten Bogor, Indonesia  
+PT NAKAFA TEKNO KREATIF, Taman Sukahati Permai Blok H-6, Sukahati, Cibinong, Kabupaten Bogor, Jawa Barat 16913, Indonesia
+
 Email: [nakafaai@gmail.com](mailto:nakafaai@gmail.com)
