+++ apps/www/app/[locale]/(app)/(shared)/(site)/(legal)/terms-of-service/en.mdx
@@ -2,7 +2,7 @@
 
 Last updated: April 24, 2026
 
-These Terms of Service (the "Terms") are a binding contract between you and PT. Nakafa Tekno Kreatif ("Nakafa," "we," "us," or "our"). These Terms govern your access to and use of our websites, applications, and related services (collectively, the "Services").
+These Terms of Service (the "Terms") are a binding contract between you and PT NAKAFA TEKNO KREATIF ("Nakafa," "we," "us," or "our"). These Terms govern your access to and use of our websites, applications, and related services (collectively, the "Services").
 
 By using the Services, you agree to these Terms. If you do not agree, do not use the Services.
 
@@ -168,7 +168,6 @@ Your continued use of the Services after the updated Terms become effective mean
 
 ## Contact
 
-PT. Nakafa Tekno Kreatif  
-Taman Sukahati Permai H6  
-Kabupaten Bogor, Indonesia  
+PT NAKAFA TEKNO KREATIF, Taman Sukahati Permai Blok H-6, Sukahati, Cibinong, Kabupaten Bogor, Jawa Barat 16913, Indonesia
+
 Email: [nakafaai@gmail.com](mailto:nakafaai@gmail.com)
