+++ apps/www/app/[locale]/(app)/(shared)/(site)/(legal)/terms-of-service/id.mdx
@@ -2,7 +2,7 @@
 
 Terakhir diperbarui: 24 April 2026
 
-Syarat dan Ketentuan ini ("Ketentuan") adalah perjanjian yang mengikat antara Anda dan PT. Nakafa Tekno Kreatif ("Nakafa", "kami"). Ketentuan ini mengatur akses dan penggunaan Anda atas situs web, aplikasi, dan layanan terkait kami (secara kolektif, "Layanan").
+Syarat dan Ketentuan ini ("Ketentuan") adalah perjanjian yang mengikat antara Anda dan PT NAKAFA TEKNO KREATIF ("Nakafa", "kami"). Ketentuan ini mengatur akses dan penggunaan Anda atas situs web, aplikasi, dan layanan terkait kami (secara kolektif, "Layanan").
 
 Dengan menggunakan Layanan, Anda menyetujui Ketentuan ini. Jika Anda tidak setuju, jangan gunakan Layanan.
 
@@ -169,7 +169,6 @@ Penggunaan Layanan secara berkelanjutan setelah Ketentuan yang diperbarui berlak
 
 ## Kontak
 
-PT. Nakafa Tekno Kreatif  
-Taman Sukahati Permai H6  
-Kabupaten Bogor, Indonesia  
+PT NAKAFA TEKNO KREATIF, Taman Sukahati Permai Blok H-6, Sukahati, Cibinong, Kabupaten Bogor, Jawa Barat 16913, Indonesia
+
 Email: [nakafaai@gmail.com](mailto:nakafaai@gmail.com)
