+++ packages/internationalization/dictionaries/de.json
@@ -91,7 +91,6 @@
     "open-in-t3-chat": "In T3 Chat öffnen",
     "open-in-gemini": "In Gemini öffnen",
     "see-on-youtube": "Auf YouTube ansehen",
-    "subscribe": "Abonnieren",
     "comments": "Kommentare",
     "anonymous": "Anonym",
     "delete": "Löschen",
