+++ packages/internationalization/dictionaries/en.json
@@ -91,7 +91,6 @@
     "open-in-t3-chat": "Open in T3 Chat",
     "open-in-gemini": "Open in Gemini",
     "see-on-youtube": "See on YouTube",
-    "subscribe": "Subscribe",
     "comments": "Comments",
     "anonymous": "Anonymous",
     "delete": "Delete",
