+++ packages/internationalization/dictionaries/id.json
@@ -91,7 +91,6 @@
     "open-in-t3-chat": "Buka di T3 Chat",
     "open-in-gemini": "Buka di Gemini",
     "see-on-youtube": "Lihat di YouTube",
-    "subscribe": "Berlangganan",
     "comments": "Komentar",
     "anonymous": "Anonim",
     "delete": "Hapus",
