+++ packages/seo/json-ld/constants.ts
@@ -1,7 +1,5 @@
-import {
-  COMPANY_IDENTITY,
-  COMPANY_SOCIAL_PROFILE_URLS,
-} from "@repo/seo/company";
+import { COMPANY_IDENTITY } from "@repo/seo/company";
+import { COMPANY_SOCIAL_PROFILE_URLS } from "@repo/seo/company-profiles";
 import type { OrganizationLeaf, Person } from "schema-dts";
 
 export const ORGANIZATION: OrganizationLeaf = {
